Alpha 0.6.0 Build
Readme changed and updated with details
Memory cards and Backup ram working and saved in the /Saves folder
PLL used for all sync of the core (With the exception of the LPSC chip for now. Now sound is completed resolve and the video is working 100% with
Reset core in the LPSC chip changed so it will sync to the correct clock edge.
Interactions done for the:
* controller selections (New Analogue Firmware required)
* H and V offsets for some games requiring the screen relocation
* PAL and NTSC selections