Alpha_0.7.5_2022-19-09
Added support for all larger games.
Change all the memory controllers so the larger roms can assess them
Moved the 68K ram to the sram chip on the pocket
Changed the interact.json with less options so it is easy for booting. Removed the dipswitches
Had to work on the reset core a bit more to sync the LPSC and the video output.